requirements.txt
imageio
numpy
pillow
tensorflow
